+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 - Changed last sync date format
 - Fixed country image vertical align
+- Fixed delete file only if exists
+- Open links on another tab
 
 ## [1.0.1]
 - Synced date now is only using javascript